plugins/YOLO_based_detector/requirements.txt
hydra-core
Pillow
opencv-python
torch
